Automatically call reset() if the user doesn't do it.
If the device isn't reset() before use, fubar happens. I'd failed to document this properly, and I'd forgotten it myself a couple times. When you're firing up the oscilloscope (for the second time in a year on the same library) to fix something that should have been automatic, you begin to question your life choices.